refactor: 리포지토리 인터페이스를 도메인으로 이동, 리포지토리는 dao로 이름 수정
