+++ src/main/java/roomescape/dao/H2ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
@@ -6,6 +6,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Name;
 import roomescape.domain.Reservation;
+import roomescape.domain.repository.ReservationRepository;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
