+++ src/main/java/roomescape/dao/H2ReservationTimeRepository.java
@@ -1,10 +1,11 @@
-package roomescape.repository;
+package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.ReservationTime;
+import roomescape.domain.repository.ReservationTimeRepository;
 
 import javax.sql.DataSource;
 import java.sql.ResultSet;
