+++ src/main/java/roomescape/dao/H2ThemeRepository.java
@@ -1,10 +1,11 @@
-package roomescape.repository;
+package roomescape.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
+import roomescape.domain.repository.ThemeRepository;
 
 import javax.sql.DataSource;
 import java.sql.ResultSet;
