+++ src/main/java/roomescape/domain/repository/ReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.domain.repository;
 
 import roomescape.domain.ReservationTime;
