+++ src/main/java/roomescape/service/ThemeService.java
@@ -5,8 +5,8 @@ import roomescape.service.dto.request.ThemeRequest;
 import roomescape.service.dto.response.ThemeResponse;
 import roomescape.domain.Reservation;
 import roomescape.domain.Theme;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ThemeRepository;
+import roomescape.domain.repository.ReservationRepository;
+import roomescape.domain.repository.ThemeRepository;
 
 import java.time.LocalDate;
 import java.util.Comparator;
